fix(scram-auth): send server nonce in second message for SCRAM auth (#46)
* fix(scram-auth): send server nonce in second message for SCRAM auth

The kafka SCRAM authentication began to fail after a fix in Kafka
for CVE-2024-56128. The Kafka SCRAM implementation failed to validate
the client sent the server nonce in the second message, which is
required by the spec and could cause a replay attack vulnerability.
The fix uncovered the issue that the Kong SCRAM implementation was
not sending the server nonce in the second message.
This PR fixes this issue.

* fix issue with delegation token generation